ebokeh/requirements.txt
torch==1.13.1
pytorch_lightning==1.9.4
piq ~= 0.7.0
scikit-image ~= 0.19.3
numpy==1.24.2
pillow ~= 9.4.0
lpips==0.1.3
torchvision~=0.15.0
tqdm~=4.65.0